🚚 Move code that defines the SetPropExtension class
Separated the code that defines the "setProp" tag into its own file ./src/template-tags/setProp.ts.
This makes it easy to define type aliases that are shared within the class.
