+++ src/utils.ts
@@ -1,0 +1,5 @@
+export function isObject(
+    value: unknown,
+): value is Record<PropertyKey, unknown> {
+    return typeof value === 'object' && value !== null;
+}
